+++ CHANGELOG.md
@@ -9,12 +9,16 @@ All notable changes will be documented in this file.
 - Command /anonymousMessage (Speak?¿)
 - Command /note
 
-## [0.0.3 - Snapshot.1] - 2018-02-14
+## [0.0.3 - Snapshot.1b] - 2018-02-14
 ### Added
 - Command /shortLink
+- Added /help text
+- Add /random command to Inline
 
 ### Changed
-- Tokens reading 
+- Tokens reading
+- /leave command for admins
+-
 
 ## [0.0.2 - Snapshot.1b] - 2018-02-13
 ### Added